+++ packages/wallet-adapter-react/src/WalletProvider.tsx
@@ -1,11 +1,4 @@
-import {
-  FC,
-  ReactNode,
-  useCallback,
-  useEffect,
-  useMemo,
-  useState,
-} from "react";
+import { FC, ReactNode, useCallback, useEffect, useRef, useState } from "react";
 import { WalletContext } from "./useWallet";
 import type {
   AccountInfo,
@@ -26,14 +19,14 @@ import type {
   AptosStandardSupportedWallet,
   AvailableWallets,
 } from "@aptos-labs/wallet-adapter-core";
-import { WalletCore } from "@aptos-labs/wallet-adapter-core";
+import { DappConfig, WalletCore } from "@aptos-labs/wallet-adapter-core";
 
 export interface AptosWalletProviderProps {
   children: ReactNode;
   plugins?: ReadonlyArray<Wallet>;
   optInWallets?: ReadonlyArray<AvailableWallets>;
   autoConnect?: boolean;
-  dappConfig?: { network: Network };
+  dappConfig?: DappConfig;
   onError?: (error: any) => void;
 }
 
@@ -74,18 +67,31 @@ export const AptosWalletAdapterProvider: FC<AptosWalletProviderProps> = ({
   // https://github.com/aptos-labs/aptos-wallet-adapter/issues/94
   const [isLoading, setIsLoading] = useState<boolean>(true);
 
-  const walletCore = useMemo(
-    () => new WalletCore(plugins ?? [], optInWallets ?? [], dappConfig),
-    []
-  );
+  const [walletCore, setWalletCore] = useState<WalletCore>();
+
   const [wallets, setWallets] = useState<
     ReadonlyArray<Wallet | AptosStandardSupportedWallet>
-  >(walletCore.wallets);
+  >(plugins ?? []);
+
+  // Initialize WalletCore on first load
+  useEffect(() => {
+    const walletCore = new WalletCore(
+      plugins ?? [],
+      optInWallets ?? [],
+      dappConfig
+    );
+    setWalletCore(walletCore);
+  }, []);
+
+  // Update initial Wallets state once WalletCore has been initialized
+  useEffect(() => {
+    setWallets(walletCore?.wallets ?? []);
+  }, [walletCore]);
 
   const connect = async (walletName: WalletName) => {
     try {
       setIsLoading(true);
-      await walletCore.connect(walletName);
+      await walletCore?.connect(walletName);
     } catch (error: any) {
       if (onError) onError(error);
       return Promise.reject(error);
@@ -96,7 +102,7 @@ export const AptosWalletAdapterProvider: FC<AptosWalletProviderProps> = ({
 
   const disconnect = async () => {
     try {
-      await walletCore.disconnect();
+      await walletCore?.disconnect();
     } catch (error) {
       if (onError) onError(error);
       return Promise.reject(error);
@@ -108,8 +114,15 @@ export const AptosWalletAdapterProvider: FC<AptosWalletProviderProps> = ({
     asFeePayer?: boolean,
     options?: InputGenerateTransactionOptions
   ): Promise<AccountAuthenticator> => {
+    if (!walletCore) {
+      throw new Error("WalletCore is not initialized");
+    }
     try {
-      return await walletCore.signTransaction(transaction, asFeePayer, options);
+      return await walletCore?.signTransaction(
+        transaction,
+        asFeePayer,
+        options
+      );
     } catch (error: any) {
       if (onError) onError(error);
       return Promise.reject(error);
@@ -119,8 +132,11 @@ export const AptosWalletAdapterProvider: FC<AptosWalletProviderProps> = ({
   const signMessage = async (
     message: SignMessagePayload
   ): Promise<SignMessageResponse> => {
+    if (!walletCore) {
+      throw new Error("WalletCore is not initialized");
+    }
     try {
-      return await walletCore.signMessage(message);
+      return await walletCore?.signMessage(message);
     } catch (error: any) {
       if (onError) onError(error);
       return Promise.reject(error);
@@ -130,8 +146,11 @@ export const AptosWalletAdapterProvider: FC<AptosWalletProviderProps> = ({
   const signMessageAndVerify = async (
     message: SignMessagePayload
   ): Promise<boolean> => {
+    if (!walletCore) {
+      throw new Error("WalletCore is not initialized");
+    }
     try {
-      return await walletCore.signMessageAndVerify(message);
+      return await walletCore?.signMessageAndVerify(message);
     } catch (error: any) {
       if (onError) onError(error);
       return Promise.reject(error);
@@ -141,8 +160,11 @@ export const AptosWalletAdapterProvider: FC<AptosWalletProviderProps> = ({
   const submitTransaction = async (
     transaction: InputSubmitTransactionData
   ): Promise<PendingTransactionResponse> => {
+    if (!walletCore) {
+      throw new Error("WalletCore is not initialized");
+    }
     try {
-      return await walletCore.submitTransaction(transaction);
+      return await walletCore?.submitTransaction(transaction);
     } catch (error: any) {
       if (onError) onError(error);
       return Promise.reject(error);
@@ -153,7 +175,7 @@ export const AptosWalletAdapterProvider: FC<AptosWalletProviderProps> = ({
     transaction: InputTransactionData
   ) => {
     try {
-      return await walletCore.signAndSubmitTransaction(transaction);
+      return await walletCore?.signAndSubmitTransaction(transaction);
     } catch (error: any) {
       if (onError) onError(error);
       return Promise.reject(error);
@@ -161,8 +183,11 @@ export const AptosWalletAdapterProvider: FC<AptosWalletProviderProps> = ({
   };
 
   const changeNetwork = async (network: Network) => {
+    if (!walletCore) {
+      throw new Error("WalletCore is not initialized");
+    }
     try {
-      return await walletCore.changeNetwork(network);
+      return await walletCore?.changeNetwork(network);
     } catch (error: any) {
       if (onError) onError(error);
       return Promise.reject(error);
@@ -182,8 +207,8 @@ export const AptosWalletAdapterProvider: FC<AptosWalletProviderProps> = ({
 
   useEffect(() => {
     if (connected) {
-      walletCore.onAccountChange();
-      walletCore.onNetworkChange();
+      walletCore?.onAccountChange();
+      walletCore?.onNetworkChange();
     }
   }, [connected]);
 
@@ -193,9 +218,9 @@ export const AptosWalletAdapterProvider: FC<AptosWalletProviderProps> = ({
       return {
         ...state,
         connected: true,
-        account: walletCore.account,
-        network: walletCore.network,
-        wallet: walletCore.wallet,
+        account: walletCore?.account || null,
+        network: walletCore?.network || null,
+        wallet: walletCore?.wallet || null,
       };
     });
   };
@@ -207,8 +232,8 @@ export const AptosWalletAdapterProvider: FC<AptosWalletProviderProps> = ({
       return {
         ...state,
         connected: false,
-        account: walletCore.account,
-        network: walletCore.network,
+        account: walletCore?.account || null,
+        network: walletCore?.network || null,
         wallet: null,
       };
     });
@@ -217,11 +242,11 @@ export const AptosWalletAdapterProvider: FC<AptosWalletProviderProps> = ({
   // Handle the adapter's account change event
   const handleAccountChange = useCallback(() => {
     if (!connected) return;
-    if (!walletCore.wallet) return;
+    if (!walletCore?.wallet) return;
     setState((state) => {
       return {
         ...state,
-        account: walletCore.account,
+        account: walletCore?.account || null,
       };
     });
   }, [connected]);
@@ -229,11 +254,11 @@ export const AptosWalletAdapterProvider: FC<AptosWalletProviderProps> = ({
   // Handle the adapter's network event
   const handleNetworkChange = useCallback(() => {
     if (!connected) return;
-    if (!walletCore.wallet) return;
+    if (!walletCore?.wallet) return;
     setState((state) => {
       return {
         ...state,
-        network: walletCore.network,
+        network: walletCore?.network || null,
       };
     });
   }, [connected]);
@@ -272,21 +297,21 @@ export const AptosWalletAdapterProvider: FC<AptosWalletProviderProps> = ({
   };
 
   useEffect(() => {
-    walletCore.on("connect", handleConnect);
-    walletCore.on("disconnect", handleDisconnect);
-    walletCore.on("accountChange", handleAccountChange);
-    walletCore.on("networkChange", handleNetworkChange);
-    walletCore.on("readyStateChange", handleReadyStateChange);
-    walletCore.on("standardWalletsAdded", handleStandardWalletsAdded);
+    walletCore?.on("connect", handleConnect);
+    walletCore?.on("disconnect", handleDisconnect);
+    walletCore?.on("accountChange", handleAccountChange);
+    walletCore?.on("networkChange", handleNetworkChange);
+    walletCore?.on("readyStateChange", handleReadyStateChange);
+    walletCore?.on("standardWalletsAdded", handleStandardWalletsAdded);
     return () => {
-      walletCore.off("connect", handleConnect);
-      walletCore.off("disconnect", handleDisconnect);
-      walletCore.off("accountChange", handleAccountChange);
-      walletCore.off("networkChange", handleNetworkChange);
-      walletCore.off("readyStateChange", handleReadyStateChange);
-      walletCore.off("standardWalletsAdded", handleStandardWalletsAdded);
+      walletCore?.off("connect", handleConnect);
+      walletCore?.off("disconnect", handleDisconnect);
+      walletCore?.off("accountChange", handleAccountChange);
+      walletCore?.off("networkChange", handleNetworkChange);
+      walletCore?.off("readyStateChange", handleReadyStateChange);
+      walletCore?.off("standardWalletsAdded", handleStandardWalletsAdded);
     };
-  }, [wallets, connected]);
+  }, [wallets, account]);
 
   return (
     <WalletContext.Provider